Create poker_database.sql
Within the SQL script, we code in the standard hand history and player id's... however, with Python, we dynamically create a table for each individual player... which will be kept in the SQL database unless explicitly deleted. This allows for us to track lifetime stats, without  wasting space keeping information on players who aren't participating in as many games or hands.